add package dto

+++ src/main/java/com/example/jobsearch/dao/EducationInfoDao.java
@@ -5,5 +5,5 @@ import com.example.jobsearch.model.EducationInfo;
 import java.util.List;
 
 public interface EducationInfoDao {
-    void saveAll(List<EducationInfo> educationInfoList, int resumeId);
+    void saveAll(List<EducationInfo> educationInfoList, long resumeId);
 }
