+++ src/main/java/com/example/jobsearch/dao/WorkExperienceInfoDao.java
@@ -5,5 +5,5 @@ import com.example.jobsearch.model.WorkExperienceInfo;
 import java.util.List;
 
 public interface WorkExperienceInfoDao {
-    void saveAll(List<WorkExperienceInfo> workExperienceInfos, int resumeId);
+    void saveAll(List<WorkExperienceInfo> workExperienceInfos, long resumeId);
 }
